+++ tasks/brace-expansion/src/constants.rs
@@ -1,0 +1,4 @@
+pub const OPEN_CHAR: char = '{';
+pub const CLOSE_CHAR: char = '}';
+pub const SEPARATOR: char = ',';
+pub const ESCAPE: char = '\\';
